backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
pydantic
python-dotenv
langchain
langchain-openai
openai